[refactor] 기존 LineTest 를 LineService 테스트로 변경 및 LineTest 생성
